Use larger "buckets" for the cache to preserve RAM
The only time the local timezone offset would change would be during
a daylight savings shift. These always happen on the half hour, so we
can assume that any request that came in with half an hour of another
request will have the same result. We can expand the cache and get
more hits, and also save a lot of RAM.